#397: add mentor authentication

+++ components/courses/app/controllers/admin/courses/mentors_controller.rb
@@ -18,14 +18,11 @@ module Admin
       end
 
       def destroy
-        react_to mentor.destroy
+        return react_to mentor.destroy if policy.allowed_to_destroy?
+        redirect_to admin_courses_season_mentors_path, alert: t('courses.flash.season-creator-destroy')
       end
 
       private
-
-      def add_season_breadcrumb
-        add_breadcrumb current_season, path: admin_courses_season_path(current_season)
-      end
 
       def add_mentor_breadcrumb
         add_breadcrumb 'mentors.plural', path: admin_courses_season_mentors_path(current_season)
@@ -41,6 +38,10 @@ module Admin
 
       def mentor
         @mentor ||= ::Courses::Mentor.find(params[:id])
+      end
+
+      def policy
+        Mentor::MentorPolicy.new(mentor, current_season)
       end
 
       def mentor_params
